requirements.txt
ultralytics>=8.2.3
opencv-python-headless>=4.5.1.48
opencv-python>=4.10.0.82
numpy>=1.26.4
pillow>=10.3.0
streamlit>=1.35.0
tempfile
time
os